Move functions to AvailableServer class.
